backend/requirements_doc_intel.txt
fastapi
uvicorn[standard]==0.23.2
pydantic
python-multipart
PyMuPDF==1.23.26
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
torch==2.1.1
transformers==4.35.2
tokenizers==0.15.0
huggingface-hub==0.17.3
google-generativeai==0.7.2
google-cloud-texttospeech==2.16.4
